💄 UI: SideSection 컴포넌트 사용하여 레이아웃 수정
