+++ src/components/MusicSearchPage/MusicInfoPinPreview.jsx
